move topbar specific utils into topbar
