+++ app/components/IfcPV.ts
@@ -14,13 +14,9 @@ export interface IfcPV {
   value?: string | number;
 }
 
-export function findPVByAddress(
-  arr: Array<IfcPV>,
-  address: string,
-): IfcPV | undefined {
-  return arr.find((b: IfcPV) => b.pvaddress == address);
-}
-
+/**
+ * Given an array of PVs, find a PV based on its human-readable name. If none found, return undefined.
+ */
 export function findPVByHumanReadableName(
   arr: Array<IfcPV>,
   human_readable_name: string,
